requirements.txt
anthropic>=0.40.0
google-generativeai>=0.8.0
youtube-transcript-api>=0.6.0
yt-dlp>=2024.1.1
python-dotenv>=1.0.0
python-slugify>=8.0.0
